+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Detection of OPcache configuration changes now takes care of web vs. WP-CLI difference.
 - Consistent reset for settings.
 - Improved translation loading.
+- [WP_CLI] `log` command have now a definition and all synopsis are up to date.
 
 ### Fixed
 - [SEC002] The password for ElasticCloud logger is in plain text in "Site Health Info" page.